+++ src/models/position.go
@@ -2,12 +2,12 @@ package models
 
 // Position model
 type Position struct {
-	X int16
-	Y int16
+	X int
+	Y int
 }
 
 // NewPosition Creates a new position
-func NewPosition(x, y int16) Position {
+func NewPosition(x, y int) Position {
 	result := Position{
 		X: x,
 		Y: y,
